+++ app/src/main/java/com/yurii/foody/authorization/AuthorizationRepository.kt
@@ -13,7 +13,7 @@ interface AuthorizationRepositoryInterface {
     suspend fun register(user: RegistrationForm): Flow<RegistrationForm>
     fun setToken(token: String)
     suspend fun getUser(id: Long): Flow<User>
-    suspend fun getUsersRoles(userId: Int? = null): Flow<Pagination<UserRole>>
+    suspend fun getUsersRoles(userId: Long? = null): Flow<Pagination<UserRole>>
 
     suspend fun setSelectedUserRole(userRole: UserRoleEnum?)
     suspend fun setUserRole(userRoleEnum: UserRoleEnum)
@@ -49,7 +49,7 @@ class AuthorizationRepository @VisibleForTesting constructor(
 
     override suspend fun getSavedUser(): User? = authDataStorage.currentUser.first()
 
-    override suspend fun getUsersRoles(userId: Int?) = Service.asFlow { api.usersService.getUsersRoles(userId) }
+    override suspend fun getUsersRoles(userId: Long?) = Service.asFlow { api.usersService.getUsersRoles(userId) }
 
     override suspend fun setSelectedUserRole(userRole: UserRoleEnum?) = authDataStorage.saveSelectedUserRole(userRole)
